# Use `@inbounds` annotations for `Dict` implementation