It seems to me that `Dict`s can be used in performance sensitive code, and ought to be as fast as possible.

I made liberty of `@propagate_inbounds` in order to factor out the "chain of custody" (so to speak) of the correctness of the indices.

EDIT: See microbenchmark below (I didn't see anything in *BaseBenchmarks.jl*...)