@KristofferC 2017-09-23
Any microbenchmarks to show?