@andyferris 2017-09-23
Give me one minute to recompile master. I did one on v0.6 that took 1.3 seconds and this branch takes 0.95 seconds... I'll post an update soon.