@andyferris 2017-09-23
Here's a function that does some general dict-y operations. I heard BenchmakTools isn't working (didn't try) so I rolled my own:

```julia
julia> function f(n)
                  keys = 1:n
                  vals1 = 2*(1:n)
                  vals2 = 3*(1:n)
                  
                  dict = Dict{Int,Int}()
                  for k in keys
                      @inbounds dict[k] = vals1[k]
                  end
                  for k in keys
                      @inbounds dict[k] = dict[k] + vals2[k]
                  end
                  out = 0
                  for k in keys
                      out += dict[k]
                  end
                  return out
              end
f (generic function with 1 method)

julia> @noinline g(n) = f(n)
g (generic function with 1 method)

julia> h(n) = for i in 1:10000; g(n); end
h (generic function with 1 method)
```

On this branch (after warmup):
```
julia> @time h(1000)
  0.944691 seconds (180.00 k allocations: 901.795 MiB, 2.59% gc time)

julia> @time h(1000)
  0.951921 seconds (180.00 k allocations: 901.795 MiB, 2.60% gc time)

julia> @time h(1000)
  0.949090 seconds (180.00 k allocations: 901.795 MiB, 2.58% gc time)
```

On master
```
julia> @time h(1000)
  1.052432 seconds (180.00 k allocations: 901.795 MiB, 2.16% gc time)

julia> @time h(1000)
  1.051774 seconds (180.00 k allocations: 901.795 MiB, 1.85% gc time)

julia> @time h(1000)
  1.065635 seconds (180.00 k allocations: 901.795 MiB, 1.84% gc time)
```

On v0.6
```
julia> @time h(1000)
  1.296142 seconds (180.00 k allocations: 901.795 MiB, 1.35% gc time)

julia> @time h(1000)
  1.288516 seconds (180.00 k allocations: 901.795 MiB, 1.33% gc time)

julia> @time h(1000)
  1.298345 seconds (180.00 k allocations: 901.795 MiB, 1.36% gc time)
```

So this PR represents a 10% improvement on this benchmark (with another 25% since v0.6... nice).