@fredrikekre 2017-10-02
> Is it normal for CircleCI 64 bit to give `ProcessExitedException`s?

Happens somewhat regularly.