@rfourquet 2017-10-02
> I didn't see anything in BaseBenchmarks.jl

cf. https://github.com/JuliaCI/BaseBenchmarks.jl/pull/109, which benchmarks some operations with `Dict`, but I need some help for fixing one test before it can be merged.