@andyferris 2017-10-02
OK, I'm sad we've regressed to stochastic testing... but I guess I should merge this then.